PIM-3832: move new attribute option import classes to the ConnectorBundle
