+++ src/Pim/Bundle/ConnectorBundle/Processor/Denormalization/ArrayConverter/Structured/AttributeOptionToStandardConverter.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\Flat;
+namespace Pim\Bundle\ConnectorBundle\Processor\Denormalization\ArrayConverter\Structured;
 
-use Pim\Bundle\BaseConnectorBundle\Exception\ArrayConversionException;
-use Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\ArrayConverter\StandardArrayConverterInterface;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Bundle\ConnectorBundle\Exception\ArrayConversionException;
+use Pim\Bundle\ConnectorBundle\Processor\Denormalization\ArrayConverter\StandardArrayConverterInterface;
 
 /**
- * Convert flat format to standard format for attribute option
+ * Convert structured format to standard format for attribute option
  *
  * @author    Nicolas Dupont <nicola@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
@@ -15,30 +14,21 @@ use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
  */
 class AttributeOptionToStandardConverter implements StandardArrayConverterInterface
 {
-    /** @var LocaleRepositoryInterface */
-    protected $localeRepository;
-
-    /**
-     * @param LocaleRepositoryInterface $localeRepository
-     */
-    public function __construct(LocaleRepositoryInterface $localeRepository)
-    {
-        $this->localeRepository = $localeRepository;
-    }
-
     /**
      * {@inheritdoc}
      *
-     * Converts flat csv array to standard structured array:
+     * Converts yaml array to standard structured array:
      *
      * Before:
      * {
      *     'attribute': 'maximum_print_size',
      *     'code': '210_x_1219_mm',
-     *     'sort_order': '2',
-     *     'label-de_DE': '210 x 1219 mm',
-     *     'label-en_US': '210 x 1219 mm',
-     *     'label-fr_FR': '210 x 1219 mm'
+     *     'sortOrder': 2,
+     *     'labels': {
+     *         'de_DE': '210 x 1219 mm',
+     *         'en_US': '210 x 1219 mm',
+     *         'fr_FR': '210 x 1219 mm'
+     *     }
      * }
      *
      * After:
@@ -56,23 +46,10 @@ class AttributeOptionToStandardConverter implements StandardArrayConverterInterf
     public function convert(array $item)
     {
         $this->validate($item);
-        $convertedItem = ['labels' => []];
-        foreach ($item as $field => $data) {
-            $isLabel = false !== strpos($field, 'label-', 0);
-            if ($isLabel) {
-                $labelTokens = explode('-', $field);
-                $labelLocale = $labelTokens[1];
-                $convertedItem['labels'][$labelLocale] = $data;
-            } else {
-                $convertedItem[$field] = $data;
-            }
-        }
-        if (!isset($convertedItem['sort_order'])) {
-            $convertedItem['sort_order'] = 1;
-        }
-        $convertedItem['sort_order'] = (int) $convertedItem['sort_order'];
+        $item['sort_order'] = $item['sortOrder'];
+        unset($item['sortOrder']);
 
-        return $convertedItem;
+        return $item;
     }
 
     /**
@@ -95,12 +72,7 @@ class AttributeOptionToStandardConverter implements StandardArrayConverterInterf
             }
         }
 
-        $authorizedFields = array_merge($requiredFields, ['sort_order']);
-        $localeCodes = $this->localeRepository->getActivatedLocaleCodes();
-        foreach ($localeCodes as $code) {
-            $authorizedFields[] = 'label-' . $code;
-        }
-
+        $authorizedFields = array_merge($requiredFields, ['sortOrder', 'labels']);
         foreach ($item as $field => $data) {
             if (!in_array($field, $authorizedFields)) {
                 throw new ArrayConversionException(
@@ -111,6 +83,12 @@ class AttributeOptionToStandardConverter implements StandardArrayConverterInterf
                     )
                 );
             }
+        }
+
+        if (isset($item['labels']) && !is_array($item['labels'])) {
+            throw new ArrayConversionException(
+                sprintf('Field "labels" must be an array, data provided is "%s"', print_r($item['labels'], true))
+            );
         }
     }
 }
