+++ src/Pim/Bundle/ConnectorBundle/Writer/Doctrine/BaseWriter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Writer\Doctrine;
+namespace Pim\Bundle\ConnectorBundle\Writer\Doctrine;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
